bot/requirements.txt
python-telegram-bot==13.7
google-api-python-client==2.85.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
psycopg2-binary==2.9.6